💬 Some small updates to the popup
The deleted line was intended to remove the line breaks in the popup,
but wasn't doing anything because it was _in front_ of the variable it
was intended to modify. Howeer, I like the line breaks, so I simply
remove this line. Also I added an extra line break in front of the
timestamp and replaced the three dots with an ellipsis.